refactor vis
